fix(tricky-pipe): run drop for serialized messages (#17)
Currently, when tricky-pipe messages are received by a `SerReceiver`,
their destructors are never run when the receiver is done using them.
This is not the case for the typed receiver, as the message is actually
moved out of the channel and dropped by the receiving code. However,
when received as serialized messages, the message is never actually
moved out of the channel, and currently, its destructor is never run.

This means that when a message type owns a heap allocation, receiving it
using the `SerReceiver` will leak that allocation. This is bad. This
issue was revealed by running the tests under Miri in 3950266.

I've fixed this by changing `SerRecvRef` to call
`core::ptr::drop_in_place` on the message when the `SerRecvRef` is
dropped. This requires adding a type-erased drop function to the
`SerVtable`. This way, destructors are run for messages that are
received as serialized data, and any heap allocations owned by those
messages are now freed.